+++ tools/check-no-diag-strings.sh
@@ -40,6 +40,7 @@ fi
 IMAGES="
 src/bsdsocket/bsdsocket.library
 src/netdev/anxnet.device
+src/netdev/anxgenet.device
 src/usergroup/usergroup.library
 src/tlslib/tls.library
 "